requirements.txt
sqlmodel
mariadb
python-dotenv

fastapi[standard]
uvicorn[standard]